@@ -1,114 +1,66 @@
 name: doclense
-description: A new Flutter project.
+description: An Indian made Open source PDF creator Android app
+publish_to: "none"
 
-# The following line prevents the package from being accidentally published to
-# pub.dev using `pub publish`. This is preferred for private packages.
-publish_to: "none" # Remove this line if you wish to publish to pub.dev
-
-# The following defines the version and build number for your application.
-# A version number is three numbers separated by dots, like 1.2.43
-# followed by an optional build number separated by a +.
-# Both the version and the builder number may be overridden in flutter
-# build by specifying --build-name and --build-number, respectively.
-# In Android, build-name is used as versionName while build-number used as versionCode.
-# Read more about Android versioning at https://developer.android.com/studio/publish/versioning
-# In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
-# Read more about iOS versioning at
-# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.0+1
+version: 1.0.1+2
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.18.5 <3.0.0"
 
 dependencies:
-  animated_splash_screen: ^1.1.0
-  camera: ^0.9.4
-  cupertino_icons: ^1.0.4
-  double_back_to_close_app: ^2.0.1
-  easy_folder_picker: ^1.1.1
   flutter:
     sdk: flutter
-  flutter_launcher_icons: ^0.9.2
+  animated_splash_screen: ^1.3.0
+  camera: ^0.10.0+4
+  cupertino_icons: ^1.0.5
+  double_back_to_close_app: ^2.1.0
+  easy_folder_picker: ^1.3.0
+  flutter_launcher_icons: ^0.10.0
   flutter_spinkit: ^5.1.0
-  flutter_svg: ^1.0.3
-  fluttertoast: ^8.0.8
+  flutter_svg: ^1.1.6
+  fluttertoast: ^8.1.1
   gallery_saver: ^2.3.2
-  hive: ^2.0.4
+  hive: ^2.2.3
   hive_flutter: ^1.1.0
-  image_cropper: ^1.4.1
-  image_picker: ^0.8.4+4
+  image_cropper: ^3.0.1
+  image_picker: ^0.8.6
   intl: ^0.17.0
-  introduction_screen: ^3.0.0
+  introduction_screen: ^3.1.1
   open_file: ^3.2.1
-  package_info_plus: ^1.4.0
-  path: ^1.8.0
-  path_provider: ^2.0.7
-  pdf: ^3.6.2
+  package_info_plus: ^1.4.2
+  path: ^1.8.2
+  path_provider: ^2.0.11
+  pdf: ^3.8.4
   pdf_viewer_plugin: ^2.0.1
-  permission_handler: ^8.3.0
+  permission_handler: ^10.2.0
   photofilters: ^3.0.1
-  printing: ^5.6.3
-  provider: ^6.0.1
-  quick_actions: ^0.6.0+7
-  rating_dialog: ^2.0.3
+  printing: ^5.9.3
+  provider: ^6.0.4
+  quick_actions: ^1.0.1
+  rating_dialog: ^2.0.4
   # share: ^2.0.4
-  share_plus: ^4.0.3
-  shared_preferences: ^2.0.8
-  source_span: ^1.8.1
-  url_launcher: ^6.0.13
-  wiredash: ^0.7.0+1
+  share_plus: ^4.5.3
+  shared_preferences: ^2.0.15
+  source_span: ^1.9.0
+  url_launcher: ^6.1.7
+  wiredash: ^1.5.0
+
 dev_dependencies:
-  build_runner: ^2.1.5
+  build_runner: ^2.3.2
   flutter_test:
     sdk: flutter
-  hive_generator: ^1.1.1
-lint: ^1.7.2
+  hive_generator: ^2.0.0
+lint: ^1.10.0
 
 flutter_icons:
   android: true
   ios: true
   image_path: "assets/app_icon.png"
 
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
-# The following section is specific to Flutter.
 flutter:
-  # The following line ensures that the Material Icons font is
-  # included with your application, so that you can use the icons in
-  # the material Icons class.
   uses-material-design: true
-
-  # To add assets to your application, add an assets section, like this:
-  # assets:
   assets:
     - assets/images/
     - assets/
     - assets/aboutPage/
     - assets/images/team/
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
-  # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware.
-  # For details regarding adding assets from package dependencies, see
-  # https://flutter.dev/assets-and-images/#from-packages
-  # To add custom fonts to your application, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts from package dependencies,
-  # see https://flutter.dev/custom-fonts/#from-packages
-dependencies_overrides:
-  collection: ^1.15.0
